requirements.txt
fastapi
uvicorn
pinecone-client
sentence-transformers
mysql-connector-python
python-dotenv
requests
